refactor: Merge Category into Product
