+++ src/test/java/gift/api/ProductControllerTest.java
@@ -3,7 +3,7 @@ package gift.api;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.global.security.JwtFilter;
 import gift.global.security.JwtUtil;
-import gift.category.entity.Category;
+import gift.product.entity.Category;
 import gift.global.error.CustomException;
 import gift.global.error.ErrorCode;
 import gift.product.api.ProductController;
